task3/requirements.txt
hazelcast-python-client==5.4.0
flask==3.0.0
requests==2.31.0

